+++ src/DiagnosticMessages.ts
@@ -612,8 +612,8 @@ export let DiagnosticMessages = {
         code: 1116,
         severity: DiagnosticSeverity.Error
     }),
-    missingExceptionVarToFollowCatch: () => ({
-        message: `Missing exception variable after 'catch' keyword`,
+    expectedExceptionVarToFollowCatch: () => ({
+        message: `Expected exception variable after 'catch' keyword`,
         code: 1117,
         severity: DiagnosticSeverity.Error
     }),
@@ -810,9 +810,17 @@ export let DiagnosticMessages = {
         code: 1153,
         severity: DiagnosticSeverity.Error
     }),
+    cannotFindTypeInCommentDoc: (name: string) => ({
+        message: `Cannot find type '${name}' in doc comment`,
+        code: 1154,
+        data: {
+            name: name
+        },
+        severity: DiagnosticSeverity.Error
+    }),
     expectedReturnStatement: () => ({
         message: `Expected return statement in function`,
-        code: 1154,
+        code: 1155,
         severity: DiagnosticSeverity.Error
     })
 };